move factory files to 'factories'
