+++ source/databricks/calculation_engine/tests/calculation/wholesale/test_fee_calculators.py
@@ -24,7 +24,7 @@ from package.codelists import (
     SettlementMethod,
 )
 from package.constants import Colname
-import tests.calculation.wholesale.prepared_fees_factory as factory
+import calculation.wholesale.factories.prepared_fees_factory as factory
 
 
 def _get_all_wholesale_metering_point_types() -> list[MeteringPointType]:
