+++ source/databricks/calculation_engine/tests/calculation/wholesale/test_subscription_calculators.py
@@ -26,7 +26,7 @@ from package.codelists import (
     SettlementMethod,
 )
 from package.constants import Colname
-import tests.calculation.wholesale.prepared_subscriptions_factory as factory
+import calculation.wholesale.factories.prepared_subscriptions_factory as factory
 
 
 class DefaultValues:
