+++ source/databricks/calculation_engine/tests/calculation/wholesale/test_tariff_calculators.py
@@ -32,7 +32,7 @@ from package.codelists import (
     WholesaleResultResolution,
 )
 from package.constants import Colname
-import tests.calculation.wholesale.prepared_tariffs_factory as factory
+import calculation.wholesale.factories.prepared_tariffs_factory as factory
 
 
 def test__calculate_tariff_price_per_ga_co_es__returns_empty_df_when_input_df_is_empty(
